+++ .ruby-version
@@ -1,1 +1,1 @@
-3.0
+2.7
