+++ Gemfile
@@ -5,8 +5,8 @@ gem 'sinatra'
 gem 'sinatra-contrib'
 
 group :test do
-  gem 'rspec'
   gem 'capybara'
+  gem 'rspec'
   gem 'rubocop', '0.79.0'
   gem 'simplecov', require: false
   gem 'simplecov-console', require: false
