+++ README.md
@@ -1,0 +1,19 @@
+### To set up the database
+
+Connect to `psql` and create the `twitter_clone` database:
+
+```
+CREATE DATABASE twitter_clone;
+```
+
+To set up the appropriate tables, connect to the database in `psql` and run the SQL scripts in the `db/migrations` folder in the given order.
+
+### To set up the test database
+
+Connect to `psql` and create the `twitter_clone_test` database:
+
+```
+CREATE DATABASE twitter_clone_test;
+```
+
+To set up the appropriate tables, connect to the database in `psql` and run the SQL scripts in the `db/migrations` folder in the given order.
