+++ spec/feature/view_tweets_spec.rb
